Remove 'email_filter_by' property from FinderSchema
This was removed in #2943 and so this commit should really have
been included there too.

A number of tests were relying on the ability to set
`email_filter_by` directly, so the blast radius on this one is
a little higher than one might imagine. That said, all of those
tests are to do with the "request edits to email alert config"
form, which is still under development, so perhaps to be expected.